WIP fix for #1755
Not sure how this is broken again, or why this is needed, but it works.

The problem seems to be a delayed error in a timer from when the
candidates are inserted before a pattern is typed.  If the DISPLAY
strings are read-only, the (text-read-only) error happens.  This fix
makes the DISPLAY strings not read-only, preventing the error.